+++ tests/e2e/image/index.js
@@ -1,0 +1,5 @@
+const { describe } = require("mocha");
+
+describe("E2E - 🖼️ Image tests", () => {
+    require("./image-fetch.test");
+});