+++ tests/e2e/index.js
@@ -1,4 +1,5 @@
 require("./app.test");
+require("./image");
 require("./user");
 require("./card");
 require("./game");